Use git:// instead of tarball for GitHub locations
It turns out that GitHub tarballs that have not been explicitly
uploaded but are simply provided by GitHub automatically for every tag,
have been confirmed to change MD5 sum.  Supposedly the tar-ball is
generated with a new tar variation now.

This has happened now for OTP-18.2.3.tar.gz The build fails 100%
guaranteed since the checksum does not match. (Download and check the
MD5 of the given tarball for 18.2.3 - the MD5 now (7fe1ad...) does
match what the previous recipe claims it to be (236741...))

Thus, an urgent fix is required.  Instead of just updating the MD5, this
change modifies the SRC_URI to instead use a git clone, so that this
issue is avoided altogether.  It does the same for the only other recipe
that seemed to use a tarball SRC_URI stemming from GitHub.

Signed-off-by: Gunnar Andersson <gandersson@genivi.org>