erlang: copy dyn_erl to erl
The `erl` is just a bash script and it works as expected when not
included in a release with include ERTS and system libs enabled.

Copying dyn_erl to erl make sure that erl will not depends on any hard
code path when running the script (variable like ROOTDIR).

See: https://github.com/erlang/rebar3/issues/2496 and
https://github.com/erlware/relx/issues/857

When making a release with `rebar3 release` rebar3 copies the dyn_erl to
erl. However, when making a release with `rebar3 tar`, the
systools:make_tar/2 overrides the `erl` with a copy of erl which is a
bash script.

This commit makes sure that the erl is always the correct binary and not
the shell script, which has variables not suitable for releasing an
application with embedded ERTS.