couchdb: fix b64url build
couchdb uses rebar2 that means rebar_port_compiler [1] will handle the C
side. That brings some additional challenges when doing cross-compile.

So, we need to use the rebar_port_compile environment variables in order
to point to the correct cross-compile sysroot. This commit uses the
pkg-config to get the version of erl_interface and add it into the
ERL_EI_LIBDIR variable. Also the correct include patch comes from pkg-config.

1:
https://github.com/rebar/rebar/blob/b6d309417c502ca243f810e5313bea36951ef038/src/rebar_port_compiler.erl#L145

Fixes: 85