mix.bbclass: prefer release output from  --path instead of tar
mix is able to write the release output into a proper directory when the
argument --path is used. This is a good way to descrease the package
time.

On the other hand, and this need further investigation, for some reason
Elixir is getting native ERTS instead the target one.

Meanwhile, let's switch to mix with --path that also works for
meta-erlang use cases.