erlang: export EGREP to make AC_EGREP_CPP works
YP/OE hardknott uses autoconf 2.71 while Erlang/OTP still uses autoconf
2.69 leading to some potential problems. One of then is about the macro
AC_EGREP_CPP which expects the EGREP variable be filled.

The issue [1] has more details about the problem. For while the
erlang.inc exports the EGREP variable to fix the problem when compiling
erlang-native that was not able to find the correct Openssl libraries in
the sysroot-native folder.

Moreover, this commit also introduce the use of '--with-ssl-zlib' to fix
a potential host contamination when building for native and native-sdk.

1: https://github.com/erlang/otp/issues/4821