rabbitmq: start rabbitmq with default rabbitmq user and group
Also, this change removes the call to rabbitmqctl because it will not work
as expected.

rabbitmqctl needs to be configured in order to find the correct erlang
cookie file. See [1] for more details.

As an alternative approach rabbitmqctl could be run as user rabbitmq but
I don't think it worth the change because it will make run rabbitmq with
`su` like: su rabbitmq -c rabbitmqctl wait <pid file>.

1: https://www.rabbitmq.com/cli.html#erlang-cookie