beamtools: add recipe
beamtools [1] provides a tarball with BEAM tools (Erlang, Elixir, rebar3,
elvis, erlfmt). It could be used by hosts where those package is not
available or the latest version is required.

The idea was borrowed from buildtools YP/OE recipe [1].

The user could use beamtools as a base to extend it if necessary.

1: https://meta-erlang.github.io/#/guides-beamtools
2:
https://docs.yoctoproject.org/ref-manual/system-requirements.html?highlight=buildtools#building-your-own-buildtools-tarball