erlang: export OPENSSL_MODULES in the wrapper
OpenSSL 3 added the concept of provider modules which are loaded from
disk.  The load path is hard-coded into the library and needs to be
relocated when running natively, so add OPENSSL_MODULES to the wrapper.

Based on commit 160ac2f136cb8df829c803848c7c47d707a908ff from OE-Core