couchdb: remove support
The couchdb build is broken because the version of mozjs supported by
couchdb is not compatible with python 3.12.x ar greater.

That brings a difficult situation where the best thing for now is to
remove couchdb and wait couchdb developers to integrate a new JS engine
[1].

In meta-erlang, patches exist in order to bring back couchdb with new JS
engine [2]. But it's not time to integrate it yet.

1: https://github.com/apache/couchdb/issues/4448 Use QuickJS as a Javascript Engine
2: https://github.com/meta-erlang/meta-erlang/pull/293 couchdb: move to main branch and remove mozjs