erlang: absorb epmd recipe
The standalone epmd recipe was difficult to follow all erlang releases.
It was stucked in OTP 25. So, the best approach is to make erlang recipe
handle epmd package generation.

A new package called erlang-epmd is provided. This package only provides
the epmd binary. There is also a package called epmd which has
erlang-epmd as runtime dependency, that package is also responsible for
installing epmd scripts (systemd, initd).

That way we can guarantee that everything will be up to date when
updating Erlang/OTP versions.

As erlang-epmd is a new package, the generate-manifest scripts has been
updated to take care of it.